# integration-suite-package/openapi.yaml
openapi: 3.0.3
info:
  title: Gravity Integration API — Pedido
  version: 1.0.0
  description: |
    API pública Gravity para integração SAP/CPI.
    Campos DDD canônicos coexistem com aliases externos (gravityOrderId, externalRef).
servers:
  - url: https://{tenant}.gravity.app/api/v1/cockpit
    variables:
      tenant:
        default: sandbox
paths:
  /oauth/token:
    post:
      summary: OAuth2 client_credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grant_type, client_id, client_secret]
              properties:
                grant_type:
                  type: string
                  enum: [client_credentials]
                client_id:
                  type: string
                client_secret:
                  type: string
      responses:
        '200':
          description: Token emitido
  /v1/orders:
    get:
      summary: Listar pedidos (paginação cursor)
      parameters:
        - name: updatedSince
          in: query
          schema:
            type: string
            format: date-time
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista paginada
    post:
      summary: Criar pedido
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Pedido criado
  /v1/orders/{id}:
    get:
      summary: Detalhe do pedido
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pedido
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    OrderCreate:
      type: object
      required: [tipo_operacao_pedido, numero_pedido]
      properties:
        tipo_operacao_pedido:
          type: string
          enum: [importacao, exportacao]
        numero_pedido:
          type: string
        externalRef:
          type: string
          description: ID do pedido no SAP (EBELN etc.)
        referencia_externa_erp_pedido:
          type: string
        id_workspace:
          type: string
        suid_importador:
          type: string
        suid_exportador:
          type: string
    Order:
      type: object
      properties:
        id_pedido:
          type: string
        gravityOrderId:
          type: string
        externalRef:
          type: string
          nullable: true
        referencia_externa_erp_pedido:
          type: string
          nullable: true
        status_pedido:
          type: string